Move block import error enum into `block_import` module
